Adjust casting in send_offset and recv_offset to avoid warnings

+++ comm/mpi.cpp
@@ -62,8 +62,8 @@ void sync_ghost_layer_loop(
     for(neigh = 0; neigh < neighs; ++neigh) {
         int send_rank = send_ranks[neigh];
         int recv_rank = recv_ranks[neigh];
-        int send_offset = send_offsets[neigh] * 3 * sizeof(double);
-        int recv_offset = recv_offsets[neigh] * 3 * sizeof(double);
+        unsigned long send_offset = send_offsets[neigh] * 3 * sizeof(double);
+        unsigned long recv_offset = recv_offsets[neigh] * 3 * sizeof(double);
 
         MPI_Irecv(&(recv_buffer[recv_offset]), recv_lengths[neigh] * 3, MPI_DOUBLE, recv_rank, 0, MPI_COMM_WORLD, &request);
         MPI_Send(&(send_buffer[send_offset]), send_lengths[neigh] * 3, MPI_DOUBLE, send_rank, 0, MPI_COMM_WORLD);
